double to float

+++ 02-electrondensity/src/Atom.cpp
@@ -19,14 +19,14 @@ static string chemSymbols[119] = {
     "Cm", "Bk", "Cf", "Es", "Fm", "Md", "No", "Lr", "Rf", "Db", "Sg", "Bh",
     "Hs", "Mt", "Ds", "Rg", "--", "--", "--", "--", "--", "--", "--"};
 
-Atom::Atom(int n, double rx, double ry, double rz) {
+Atom::Atom(int n, float rx, float ry, float rz) {
   zatom = n;
   mass = setAtomicMass(n);
   symb = setSymbol(n);
   coor = Rvector(rx, ry, rz);
 }
 
-Atom::Atom(int n, double *r) {
+Atom::Atom(int n, float *r) {
   zatom = n;
   mass = setAtomicMass(n);
   symb = setSymbol(n);
@@ -40,14 +40,14 @@ Atom::Atom(int n, Rvector r) {
   coor = Rvector(r);
 }
 
-Atom::Atom(string st, double rx, double ry, double rz) {
+Atom::Atom(string st, float rx, float ry, float rz) {
   zatom = setAtomicNumberfromSymbol(st);
   mass = setAtomicMass(zatom);
   symb = st;
   coor = Rvector(rx, ry, rz);
 }
 
-Atom::Atom(string st, double *r) {
+Atom::Atom(string st, float *r) {
   zatom = setAtomicNumberfromSymbol(st);
   mass = setAtomicMass(zatom);
   symb = st;
@@ -61,14 +61,14 @@ Atom::Atom(string st, Rvector r) {
   coor = Rvector(r);
 }
 
-Atom::Atom(const char *st, double rx, double ry, double rz) {
+Atom::Atom(const char *st, float rx, float ry, float rz) {
   zatom = setAtomicNumberfromSymbol(string(st));
   mass = setAtomicMass(zatom);
   symb = string(st);
   coor = Rvector(rx, ry, rz);
 }
 
-Atom::Atom(const char *st, double *r) {
+Atom::Atom(const char *st, float *r) {
   zatom = setAtomicNumberfromSymbol(string(st));
   mass = setAtomicMass(zatom);
   symb = string(st);
@@ -103,8 +103,8 @@ int Atom::setAtomicNumberfromSymbol(string symbol) {
   return 0;
 }
 
-double Atom::setAtomicMass(int idx) {
-  static double chemAtomicMass[119] = {
+float Atom::setAtomicMass(int idx) {
+  static float chemAtomicMass[119] = {
       0.0,      1.008,    4.0026,   6.9410,   9.0122,   10.8110,  12.011,
       14.0067,  15.9994,  18.9984,  20.1797,  22.9898,  24.3050,  26.9815,
       28.0855,  30.9737,  32.0650,  35.4530,  39.9480,  39.0983,  40.0780,
@@ -130,19 +130,19 @@ double Atom::setAtomicMass(int idx) {
 
 Rvector Atom::getCoors() { return coor; }
 
-double Atom::get_x() { return coor[0]; }
+float Atom::get_x() { return coor[0]; }
 
-double Atom::get_y() { return coor[1]; }
+float Atom::get_y() { return coor[1]; }
 
-double Atom::get_z() { return coor[2]; }
+float Atom::get_z() { return coor[2]; }
 
-double Atom::getMass() { return mass; }
+float Atom::getMass() { return mass; }
 
 string Atom::getSymbol() { return symb; }
 
 int Atom::get_atnum() { return zatom; }
 
-double Atom::get_charge() { return (double)zatom; }
+float Atom::get_charge() { return (float)zatom; }
 
 ostream &operator<<(ostream &o, const Atom &a) {
   o << fixed << setw(3) << a.symb << setw(6) << a.zatom
