+++ 02-electrondensity/src/Field.hpp
@@ -10,30 +10,30 @@
 
 class Field {
 public:
-  Field(Wavefunction &wf, double rmin, double delta);
+  Field(Wavefunction &wf, float rmin, float delta);
 
-  double orbital(Rvector r, int i);
-  double gaussiana(Rvector r, int mu);
-  double Density(Rvector r);
+  float orbital(Rvector r, int i);
+  float gaussiana(Rvector r, int mu);
+  float Density(Rvector r);
   void evalDensity();
   void evalDensity2();
   void evalDensity2D();
 
   void evalDensity_sycl();
   void evalDensity_sycl2();
-  static SYCL_EXTERNAL double Density(int, int, const int *, const int *,
-                                           const double *, const double *,
-                                           const double *, const double *,
-                                           const double *);
+  static SYCL_EXTERNAL float Density(int, int, const int *, const int *,
+                                           const float *, const float *,
+                                           const float *, const float *,
+                                           const float *);
 
   void spherical(std::string fname);
 
   void dumpXYZ(std::string filename);
 
-//  void dumpCube(double, double, double, double, int, int, int, vector<double>,
+//  void dumpCube(float, float, float, float, int, int, int, vector<float>,
 //                std::string filename);
-  void dumpCube(double xmin, double ymin, double zmin, double delta,
-                       int nx, int ny, int nz, vector<double> field,
+  void dumpCube(float xmin, float ymin, float zmin, float delta,
+                       int nx, int ny, int nz, vector<float> field,
                        std::string filename) {
     std::ofstream fout(filename);
     if (fout.is_open()) {
@@ -96,8 +96,8 @@ public:
 
 private:
   Wavefunction &wf;
-  double xmin, ymin, zmin;
-  double delta;
+  float xmin, ymin, zmin;
+  float delta;
   int npoints_x;
   int npoints_y;
   int npoints_z;
